# src/main/java/com/acubillos/rickandmortychallenge/openapi/api/CharactersApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.27).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.acubillos.rickandmortychallenge.openapi.api;

import com.acubillos.rickandmortychallenge.openapi.model.CharactersOutput;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2022-06-22T22:50:32.230Z")

@Validated
@Api(value = "characters", description = "the characters API")
@RequestMapping(value = "/api/v1")
public interface CharactersApi {

    @ApiOperation(value = "get a character", nickname = "charactersIdGet", notes = "use an id to get a character", response = CharactersOutput.class, tags={ "characters", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = CharactersOutput.class),
        @ApiResponse(code = 400, message = "Invalid ID supplied"),
        @ApiResponse(code = 404, message = "Character not found"),
        @ApiResponse(code = 500, message = "Internal server error") })
    @RequestMapping(value = "/characters/{id}",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.GET)
    ResponseEntity<CharactersOutput> charactersIdGet(@ApiParam(value = "",required=true) @PathVariable("id") String id);

}