# src/main/java/com/acubillos/rickandmortychallenge/openapi/api/LocationsApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.27).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.acubillos.rickandmortychallenge.openapi.api;

import com.acubillos.rickandmortychallenge.openapi.model.LocationOutput;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2022-06-22T22:50:32.230Z")

@Validated
@Api(value = "locations", description = "the locations API")
@RequestMapping(value = "/api/v1")
public interface LocationsApi {

    @ApiOperation(value = "get a location", nickname = "locationsIdGet", notes = "use an id to get a locations", response = LocationOutput.class, tags={ "locations", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = LocationOutput.class),
        @ApiResponse(code = 400, message = "Invalid ID supplied"),
        @ApiResponse(code = 404, message = "Character not found"),
        @ApiResponse(code = 500, message = "Internal server error") })
    @RequestMapping(value = "/locations/{id}",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.GET)
    ResponseEntity<LocationOutput> locationsIdGet(@ApiParam(value = "",required=true) @PathVariable("id") String id);

}