# src/main/resources/api.yaml
swagger: "2.0"
info:
  description: "rick and morty api"
  version: "1.0.0"
  title: "Rick and Morty Clean Connector"
host: "localhost"
basePath: "/api/v1"
tags:
  - name: "characters"
    description: "Everything about characters"
  - name: "locations"
    description: "Everything about locations"
  - name: "episodes"
    description: "Everything about episodes"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
  /characters/{id}:
    get:
      tags:
        - "characters"
      summary: "get a character"
      description: "use an id to get a character"
      parameters:
        - in: "path"
          name: "id"
          required: true
          type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/charactersOutput"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Character not found"
        "500":
          description: "Internal server error"

  /locations/{id}:
    get:
      tags:
        - "locations"
      summary: "get a location"
      description: "use an id to get a locations"
      parameters:
        - in: "path"
          name: "id"
          required: true
          type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/locationOutput"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Character not found"
        "500":
          description: "Internal server error"

definitions:
  charactersOutput:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      status:
        type: "string"
      species:
        type: "string"
      type:
        type: "string"
      episode_count:
        type: "integer"
      origin:
        $ref: "#/definitions/locationOutput"
  locationOutput:
    type: "object"
    properties:
      name:
        type: "string"
      url:
        type: "string"
      dimension:
        type: "string"
      residents:
        type: "array"
        items:
          type: "string"
  errors:
    type: "object"
    required:
      - "message"
    properties:
      message:
        type: "string"